Session auto source plugin
This is a quick lil script that will attempt to detect when you change
directories in vim.  If there is a `Session.vim` in the directory you
`cd`-d into, it will ask if you'd like to source that session file.

If I source the session file within the `DirChanged` event, then I get
some pretty big bugs in Vim - `filetype`s won't be set on all the opened
files.

If I have the delay too low, then it triggers a segfault in vim.